Update cookbook_0.6
copy udevil.conf to /etc, comment out 'disable getty@tty1' as you can no longer connect on the console if there is a problem. THis script as it is today does not connect to the lan on eth0. I believe we need to explicitly add a line 'enable netctl eth0' to get it to connect at boot.